requirements.txt
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
bcrypt==4.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dnspython==2.7.0
docopt==0.6.2
email_validator==2.2.0
Faker==37.5.3
fastapi==0.115.6
fonttools==4.55.3
greenlet==3.2.3
h11==0.14.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.0
mysql==0.0.3
mysql-connector-python==9.1.0
mysqlclient==2.2.7
numpy==2.2.1
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
protobuf==4.25.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
pydantic==2.10.4
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2024.2
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.3.0
uvicorn==0.34.0
yarg==0.1.10
